Proof of concept obsp, varp
This commit adds two new properties for anndata. These PairwiseArrays (a subclass of AlignedMapping) where the first two axes of their array-like values have the same shape as one dimension of their parent AnnData. These would be good for graph annotations, and would mean we wouldn't have to bother subsetting any part of uns (or copying it for that matter).

I'm labelling this as proof of concept, because there is no support for reading or writing, and no tests. My current throughts are to add this feature later after reading and writing has become more structured.