Reduced code duplication for AlignedMappings
This moves implementations of `_view` and `copy` to the parent class, since all the code was pretty close to the same anyways.
This required adding references to the relevant view and actual classes to each of the base classes. I think the benefits from reduction in duplication outweigh the non-pythonic way of doing circular references.