Address #188, and test it a bunch
There were a couple things going on here. The main issue was that if we tried to write from a backed object to a new file, it wouldn't work if the backed matrix was sparse. This was due to (I think) some logic in writing and (definitley) backed sparse matrices attempting to be copied when the whole matrix is sliced (i.e. `mtx[:]`).