Fixed recarray string columns reading w h5py v2.10
In h5py v2.10, fixed length strings in compound datasets returns a slightly different dtype that it did previously.
Now for fixed length strings, the returned array `.dtype.descr` has an annotation mentioning what kind of string it is.
This broke older reading code, which was sniffing for the dtype using the type of that field.

This commit checks for this case, though it's very possible this is

1. A breaking change for h5py, since this is external facing API
2. There is a better way for us to be checking dtypes