Fix unicode io for zarr
This makes it so we will always read an object array from disk. However, this writes fixed length unicode data to disk for zarr in structured arrays. In future, we'll write vlength arrays for text data, but that doesn't work at the moment.

One difficulty with implementing and testing this is that you can't compare a structured array with variable length strings to a structured array of fixed length strings in numpy. This could use a work-around.