Make inner and outer joins apply to obsm
This means that if `join="inner"`, only keys which are in all values of obsm will joined. In addition, the size of the resulting arrays will be the intersect of the column names (for dataframes) or the smallest number of columns for arrays.

This code is kinda messy, and should probably be replaced with a more elegant solution later. Pretty sure it works for now though.