Fix akward array view test
The test didn't catch that when modifying an awkward array in a
AnnDataView, the changes did not persist within that view.

The issue was that I had been retreiving the awkward array
from the view once, and then did tests on it. Instead, to properly
test this, I need to retreive the Awkward array from the AnnData view
every time. It turns out doing so gives me a fresh copy of the origina
awkward array and modifications only affect that copy - not the original data.